Refactor part of solstice into dev.equo.ide
