+++ solstice/src/main/java/dev/equo/ide/IdeMainUi.java
@@ -11,9 +11,11 @@
  * Contributors:
  *     EquoTech, Inc. - initial API and implementation
  *******************************************************************************/
-package dev.equo.solstice;
+package dev.equo.ide;
 
+import dev.equo.solstice.Solstice;
 import java.util.HashMap;
+import java.util.Hashtable;
 import java.util.Map;
 import org.eclipse.core.runtime.internal.adaptor.EclipseAppLauncher;
 import org.eclipse.equinox.app.IApplicationContext;
@@ -40,7 +42,8 @@ class IdeMainUi {
 		var appDescriptor = osgiShim.getService(appServices.iterator().next());
 
 		var appLauncher = new EclipseAppLauncher(osgiShim, false, false, null, null);
-		osgiShim.registerService(ApplicationLauncher.class, appLauncher, Dictionaries.empty());
+		var emptyDict = new Hashtable<String, Object>();
+		osgiShim.registerService(ApplicationLauncher.class, appLauncher, emptyDict);
 
 		Map<String, Object> appProps = new HashMap<>();
 		appProps.put(IApplicationContext.APPLICATION_ARGS, new String[] {});
