+++ solstice/src/test/java/dev/equo/ide/IdeMainBuildPluginTest.java
@@ -11,7 +11,7 @@
  * Contributors:
  *     EquoTech, Inc. - initial API and implementation
  *******************************************************************************/
-package dev.equo.solstice;
+package dev.equo.ide;
 
 import java.io.IOException;
 import org.osgi.framework.BundleException;
